refactor: share drizzle repositories via packages/db
Extract author / character / ip / preset / project / source / tag
repositories and media-search utilities into @solid-imager/db as
factories. server / tauri now keep thin wrappers that inject their
own DrizzleExecutor provider. Removes the duplicate server-side
authors-repository.ts and updates the server-tauri parity wiki.
