+++ apps/server/src/infrastructure/api/routers/authors-router.ts
@@ -1,6 +1,10 @@
 import { os } from "@orpc/server";
-import { AuthorsRepository } from "~/infrastructure/repositories/authors-repository";
+import { AuthorService } from "~/application/services/author-service";
 
 export const authorsRouter = os.router({
-	list: os.handler(async () => await AuthorsRepository.list()),
+	list: os.handler(async () =>
+		(await AuthorService.getAllAuthors()).toSorted((left, right) =>
+			right.name.localeCompare(left.name),
+		),
+	),
 });
