+++ packages/db/src/index.ts
@@ -1,3 +1,10 @@
 export * from "./migrations";
+export * from "./repositories/author-repository";
+export * from "./repositories/character-repository";
+export * from "./repositories/ip-repository";
 export * from "./repositories/media-search";
+export * from "./repositories/preset-repository";
+export * from "./repositories/project-repository";
+export * from "./repositories/source-repository";
+export * from "./repositories/tag-repository";
 export * from "./schema";
